+++ src/etoolbox/datazip/__init__.py
@@ -1,6 +1,5 @@
 """Core :class:`.DataZip` object plus mixins and wrappers."""
 from etoolbox.datazip.core import DataZip
 from etoolbox.datazip.mixin import IOMixin
-from etoolbox.datazip.wrapper import IOWrapper
 
-__all__ = ["DataZip", "IOMixin", "IOWrapper"]
+__all__ = ["DataZip", "IOMixin"]